SBERDOMA-940 - remove data mapper global reference & add as props
