+++ apps/condo/pages/reports/detail/report_by_tickets/index.tsx
@@ -14,7 +14,7 @@ import {
     Tabs,
     Divider,
     Select,
-    Tooltip, Form, notification,
+    Tooltip, Form, notification, TableColumnsType,
 } from 'antd'
 import { useOrganization } from '@core/next/organization'
 import get from 'lodash/get'
@@ -31,7 +31,7 @@ import RadioGroupWithIcon, { radioButtonBorderlessCss } from '@condo/domains/com
 import { useRouter } from 'next/router'
 import qs from 'qs'
 import DateRangePicker from '@condo/domains/common/components/DateRangePicker'
-import { ViewModeTypes, TicketSelectTypes } from '@condo/domains/ticket/components/TicketChart'
+import TicketChart, { ViewModeTypes, TicketSelectTypes } from '@condo/domains/ticket/components/TicketChart'
 import { filterToQuery, specificationTypes, ticketAnalyticsPageFilters } from '@condo/domains/ticket/utils/helpers'
 import { GraphQlSearchInput } from '@condo/domains/common/components/GraphQlSearchInput'
 import { searchProperty } from '@condo/domains/ticket/utils/clientSchema/search'
@@ -232,6 +232,7 @@ const TicketAnalyticsPage: ITicketAnalyticsPage = () => {
     const userOrganizationId = get(userOrganization, ['organization', 'id'])
 
     const filtersRef = useRef(null)
+    const mapperInstanceRef = useRef(null)
     const [groupTicketsBy, setGroupTicketsBy] = useState<groupTicketsByTypes>('status')
     const [viewMode, setViewMode] = useState<ViewModeTypes>('line')
     const [analyticsData, setAnalyticsData] = useState(null)
@@ -256,6 +257,119 @@ const TicketAnalyticsPage: ITicketAnalyticsPage = () => {
         },
     })
     const getAnalyticsData = () => {
+        if (mapperInstanceRef.current === null) {
+            mapperInstanceRef.current = new TicketChart({
+                line: {
+                    chart: (viewMode, data) => {
+                        const axisLabels = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+                        const legend = Object.keys(data)
+                        const series = []
+                        Object.entries(data).map(([groupBy, dataObj]) => {
+                            series.push({
+                                name: groupBy,
+                                type: viewMode,
+                                symbol: 'none',
+                                stack: groupBy,
+                                data: Object.values(dataObj),
+                                emphasis: {
+                                    focus: 'none',
+                                    blurScope: 'none',
+                                },
+                            })
+                        })
+                        const axisData = { yAxis: { type: 'value', data: null }, xAxis: { type: 'category', data: axisLabels } }
+                        const tooltip = { trigger: 'axis', axisPointer: { type: 'line' } }
+                        return { series, legend, axisData, tooltip }
+                    },
+                    table: (viewMode, data, restOptions) => {
+                        const dataSource = []
+                        const { translations, filters } = restOptions
+                        const tableColumns: TableColumnsType = [
+                            { title: translations['address'], dataIndex: 'address', key: 'address', sorter: (a, b) => a['address'] - b['address'] },
+                            {
+                                title: translations['date'],
+                                dataIndex: 'date',
+                                key: 'date',
+                                defaultSortOrder: 'descend',
+                                sorter: (a, b) => moment(a['date'], DATE_DISPLAY_FORMAT).unix() - moment(b['date'], DATE_DISPLAY_FORMAT).unix(),
+                            },
+                            ...Object.entries(data).map(([key]) => ({ title: key, dataIndex: key, key, sorter: (a, b) =>a[key] - b[key] })),
+                        ]
+                        const uniqueDates = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+                        uniqueDates.forEach((date, key) => {
+                            const restTableColumns = {}
+                            Object.keys(data).forEach(ticketType => (restTableColumns[ticketType] = data[ticketType][date]))
+                            let address = translations['allAddresses']
+                            const addressList = get(filters, 'addresses')
+                            if (addressList && addressList.length) {
+                                address = addressList.join(', ')
+                            }
+                            dataSource.push({ key, address, date, ...restTableColumns })
+                        })
+                        return { dataSource, tableColumns }
+                    },
+                },
+                bar: {
+                    chart: (viewMode, data) => {
+                        const series = []
+                        const axisLabels = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+                        const legend = Object.keys(data)
+                        Object.entries(data).map(([groupBy, dataObj]) => {
+                            series.push({
+                                name: groupBy,
+                                type: viewMode,
+                                symbol: 'none',
+                                stack: 'total',
+                                data: Object.values(dataObj),
+                                emphasis: {
+                                    focus: 'self',
+                                    blurScope: 'self',
+                                },
+                            })
+                        })
+                        const axisData = { yAxis: { type: 'category', data: axisLabels }, xAxis: { type: 'value', data: null } }
+                        const tooltip = { trigger: 'item', axisPointer: { type: 'line' } }
+                        return { series, legend, axisData, tooltip }
+                    },
+                    table: (viewMode, data, restOptions) => {
+                        const { translations, filters } = restOptions
+                        const dataSource = []
+                        const tableColumns: TableColumnsType = [
+                            { title: translations['address'], dataIndex: 'address', key: 'address', sorter: (a, b) => a['address'] - b['address'] },
+                            ...Object.entries(data).map(([key]) => ({ title: key, dataIndex: key, key, sorter: (a, b) => a[key] - b[key] })),
+                        ]
+                        const restTableColumns = {}
+                        const addressList = get(filters, 'addresses')
+                        const aggregateSummary = addressList !== undefined && addressList.length === 0
+                        if (aggregateSummary) {
+                            Object.entries(data).forEach((rowEntry) => {
+                                const [ticketType, dataObj] = rowEntry
+                                const counts = Object.values(dataObj) as number[]
+                                restTableColumns[ticketType] = counts.reduce((a, b) => a + b, 0)
+                            })
+                            dataSource.push({
+                                key: 0,
+                                address: translations['allAddresses'],
+                                ...restTableColumns,
+                            })
+                        } else {
+                            addressList.forEach((address, key) => {
+                                const tableRow = { key, address }
+                                Object.entries(data).forEach(rowEntry => {
+                                    const [ticketType, dataObj] = rowEntry
+                                    const counts = Object.entries(dataObj)
+                                        .filter(obj => obj[0] === address).map(e => e[1]) as number[]
+                                    tableRow[ticketType] = counts.reduce((a, b) => a + b, 0)
+                                })
+                                dataSource.push(tableRow)
+                            })
+                        }
+                        return { dataSource, tableColumns }
+                    },
+
+                },
+            })
+        }
         if (filtersRef.current !== null) {
             setLoading(true)
             const { AND, groupBy } = filterToQuery(filtersRef.current, viewMode, ticketType)
@@ -358,6 +472,7 @@ const TicketAnalyticsPage: ITicketAnalyticsPage = () => {
                                     animationEnabled: true,
                                     chartOptions: { renderer: 'svg', height: viewMode === 'line' ? 440 : 'auto' },
                                 }}
+                                mapperInstance={mapperInstanceRef.current}
                             >
                                 <Select
                                     value={ticketType}
@@ -381,6 +496,7 @@ const TicketAnalyticsPage: ITicketAnalyticsPage = () => {
                                 loading={loading}
                                 viewMode={viewMode}
                                 filters={filtersRef.current}
+                                mapperInstance={mapperInstanceRef.current}
                             />
                         ), [analyticsData, loading, viewMode])}
                     </Col>
