+++ apps/condo/pages/reports/detail/report_by_tickets/pdf.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef, useState } from 'react'
 import { OrganizationRequired } from '@condo/domains/organization/components/OrganizationRequired'
-import { Col, notification, Row, Typography } from 'antd'
+import { Col, notification, Row, TableColumnsType, Typography } from 'antd'
 import dynamic from 'next/dynamic'
 import { useIntl } from '@core/next/intl'
 import TicketChartView from '@condo/domains/ticket/components/analytics/TicketChartView'
@@ -14,9 +14,10 @@ import { createPdfWithPageBreaks } from '@condo/domains/common/utils/pdf'
 import moment from 'moment'
 import { filterToQuery } from '@condo/domains/ticket/utils/helpers'
 import { Loader } from '@condo/domains/common/components/Loader'
-import { PDF_REPORT_WIDTH } from '@condo/domains/ticket/constants/common'
+import { DATE_DISPLAY_FORMAT, PDF_REPORT_WIDTH } from '@condo/domains/ticket/constants/common'
 import { Logo } from '@condo/domains/common/components/Logo'
 import { colors } from '@condo/domains/common/constants/style'
+import TicketChart from '@condo/domains/ticket/components/TicketChart'
 
 const PdfView = () => {
     const intl = useIntl()
@@ -30,6 +31,7 @@ const PdfView = () => {
     const LoadingTip = intl.formatMessage({ id: 'pages.condo.analytics.TicketAnalyticsPage.PDF.LoadingTip' })
     const containerRef = useRef<null | HTMLDivElement>(null)
     const queryParamsRef = useRef(null)
+    const mapperInstanceRef = useRef(null)
 
     const [data, setData] = useState(null)
     const [loading, setLoading] = useState(true)
@@ -75,6 +77,119 @@ const PdfView = () => {
     }, [userOrganizationId])
 
     useEffect(() => {
+        if (mapperInstanceRef.current === null) {
+            mapperInstanceRef.current = new TicketChart({
+                line: {
+                    chart: (viewMode, data) => {
+                        const axisLabels = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+                        const legend = Object.keys(data)
+                        const series = []
+                        Object.entries(data).map(([groupBy, dataObj]) => {
+                            series.push({
+                                name: groupBy,
+                                type: viewMode,
+                                symbol: 'none',
+                                stack: groupBy,
+                                data: Object.values(dataObj),
+                                emphasis: {
+                                    focus: 'none',
+                                    blurScope: 'none',
+                                },
+                            })
+                        })
+                        const axisData = { yAxis: { type: 'value', data: null }, xAxis: { type: 'category', data: axisLabels } }
+                        const tooltip = { trigger: 'axis', axisPointer: { type: 'line' } }
+                        return { series, legend, axisData, tooltip }
+                    },
+                    table: (viewMode, data, restOptions) => {
+                        const dataSource = []
+                        const { translations, filters } = restOptions
+                        const tableColumns: TableColumnsType = [
+                            { title: translations['address'], dataIndex: 'address', key: 'address', sorter: (a, b) => a['address'] - b['address'] },
+                            {
+                                title: translations['date'],
+                                dataIndex: 'date',
+                                key: 'date',
+                                defaultSortOrder: 'descend',
+                                sorter: (a, b) => moment(a['date'], DATE_DISPLAY_FORMAT).unix() - moment(b['date'], DATE_DISPLAY_FORMAT).unix(),
+                            },
+                            ...Object.entries(data).map(([key]) => ({ title: key, dataIndex: key, key, sorter: (a, b) =>a[key] - b[key] })),
+                        ]
+                        const uniqueDates = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+                        uniqueDates.forEach((date, key) => {
+                            const restTableColumns = {}
+                            Object.keys(data).forEach(ticketType => (restTableColumns[ticketType] = data[ticketType][date]))
+                            let address = translations['allAddresses']
+                            const addressList = get(filters, 'addresses')
+                            if (addressList && addressList.length) {
+                                address = addressList.join(', ')
+                            }
+                            dataSource.push({ key, address, date, ...restTableColumns })
+                        })
+                        return { dataSource, tableColumns }
+                    },
+                },
+                bar: {
+                    chart: (viewMode, data) => {
+                        const series = []
+                        const axisLabels = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+                        const legend = Object.keys(data)
+                        Object.entries(data).map(([groupBy, dataObj]) => {
+                            series.push({
+                                name: groupBy,
+                                type: viewMode,
+                                symbol: 'none',
+                                stack: 'total',
+                                data: Object.values(dataObj),
+                                emphasis: {
+                                    focus: 'self',
+                                    blurScope: 'self',
+                                },
+                            })
+                        })
+                        const axisData = { yAxis: { type: 'category', data: axisLabels }, xAxis: { type: 'value', data: null } }
+                        const tooltip = { trigger: 'item', axisPointer: { type: 'line' } }
+                        return { series, legend, axisData, tooltip }
+                    },
+                    table: (viewMode, data, restOptions) => {
+                        const { translations, filters } = restOptions
+                        const dataSource = []
+                        const tableColumns: TableColumnsType = [
+                            { title: translations['address'], dataIndex: 'address', key: 'address', sorter: (a, b) => a['address'] - b['address'] },
+                            ...Object.entries(data).map(([key]) => ({ title: key, dataIndex: key, key, sorter: (a, b) => a[key] - b[key] })),
+                        ]
+                        const restTableColumns = {}
+                        const addressList = get(filters, 'addresses')
+                        const aggregateSummary = addressList !== undefined && addressList.length === 0
+                        if (aggregateSummary) {
+                            Object.entries(data).forEach((rowEntry) => {
+                                const [ticketType, dataObj] = rowEntry
+                                const counts = Object.values(dataObj) as number[]
+                                restTableColumns[ticketType] = counts.reduce((a, b) => a + b, 0)
+                            })
+                            dataSource.push({
+                                key: 0,
+                                address: translations['allAddresses'],
+                                ...restTableColumns,
+                            })
+                        } else {
+                            addressList.forEach((address, key) => {
+                                const tableRow = { key, address }
+                                Object.entries(data).forEach(rowEntry => {
+                                    const [ticketType, dataObj] = rowEntry
+                                    const counts = Object.entries(dataObj)
+                                        .filter(obj => obj[0] === address).map(e => e[1]) as number[]
+                                    tableRow[ticketType] = counts.reduce((a, b) => a + b, 0)
+                                })
+                                dataSource.push(tableRow)
+                            })
+                        }
+                        return { dataSource, tableColumns }
+                    },
+
+                },
+            })
+        }
         if (!loading && data !== null) {
             createPdfWithPageBreaks({ element: containerRef.current, fileName: 'analytics_result.pdf' })
                 .catch((e) => {
@@ -110,6 +225,7 @@ const PdfView = () => {
                     data={data}
                     viewMode={viewMode}
                     onChartReady={() => setLoading(false)}
+                    mapperInstance={mapperInstanceRef.current}
                     chartConfig={{
                         animationEnabled: false,
                         chartOptions: { renderer: 'svg' },
@@ -117,11 +233,16 @@ const PdfView = () => {
                 />
             </Col>
             <Col flex={1} >
-                <TicketListView data={data} viewMode={viewMode} filters={{
-                    range: [dateFrom, dateTo],
-                    addressList: addressListParsed,
-                    specification: specification,
-                }} />
+                <TicketListView
+                    mapperInstance={mapperInstanceRef.current}
+                    data={data}
+                    viewMode={viewMode}
+                    filters={{
+                        range: [dateFrom, dateTo],
+                        addressList: addressListParsed,
+                        specification: specification,
+                    }}
+                />
             </Col>
         </Row>
     </>
